Favor `abi.encodeCall` when encoding function calls (#34)
Style PR that replaces usage of `abi.encodeWithSelector` with
`abi.encodeCall`.

This is done because `encodeWithSelector` doesn't check that the input
values are consistent with the function at the called selector, while
`encodeCall` does so at compile time. This can end up being handy when
we refactor tests after changing the type or input values of a function.

Concretely, if we have the following lines in a test (note the incorrect
input for `balanceOf`):
- `abi.encodeWithSelector(IERC20.balanceOf.selector, user, someNumber,
someByteArray)` ⇒ tests run but at the point of the call would likely
fail (or the mock wouldn't be called)
- `abi.encodeCall(IERC20.balanceOf, (user, someNumber, someByteArray))`
⇒ compile-time error.

Note that this can't be done for _error_ or _event_ selectors, and
there's no corresponding `abi.encodeEvent` unfortunately, as far as I
know.

### How to test

CI still works.