Forked test showcasing the use of the helper interface (#92)
This PR adds a minimal test showing how the helper contract can be used
to generate valid AMM orders.

The helper contract hardcodes the contracts used in the various chains,
so I have no choice but using a forked test.

I used the legacy path because there's no real CoW AMM with meaningful
funds that isn't legacy.

Note that this test won't work once we migrated funds from the AMM used
in the test. This doesn't matter as this test will have lost its utility
at that point.

### How to test

CI.
Look at the code to determine that we test everything we want from the
interface.